src/workflow/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Consult description"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-aiplatform>=1.120.0",
    "kfp>=2.14.6",
]
